refactor : 첫번째 피드백 반영 매직넘버 제거, bigDecimal 사용, map 사용, 객체가 하는일 재분배, 변수이름 수정

+++ src/test/java/lotto/LottoPriceTest.java
@@ -20,7 +20,7 @@ class LottoPriceTest {
     @DisplayName("로또 1장의 금액은 1000원이다.")
     @Test
     void lottoPriceTest() {
-        assertThat(lottoPrice.getLottoPrice()).isEqualTo(1000);
+        assertThat(lottoPrice.getLottoPrice().intValue()).isEqualTo(1000);
     }
 
     @DisplayName("로또의 갯수는 구매금액 / 로또금액 이다.")
